pyproject.toml
[tool.poetry]
name = "shinywaffle"
version = "0.1.0"
description = ""
authors = ["Julian Wachholz <julian@wachholz.ch>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0"
psycopg2-binary = "^2.8"
argon2-cffi = "^19.2"
django-environ = "^0.4.5"
django-mptt = "^0.10.0"
django-countries = "^5.5"
djangorestframework = "^3.10"
python-dateutil = "^2.8"
django-filter = "^2.2"
django-rest-knox = "^4.1.0"
django-rest-auth = "^0.9.5"
uwsgi = "^2.0.18"
tablib = {extras = ["xlsx"], version = "^1.0.0"}
python-magic = "^0.4.15"
sentry-sdk = "^0.14.0"
sparkpost = "^1.3.6"
django-autocompletefilter = "^0.0.6"
whitenoise = "^5.0.1"
django-money = "^1.0"
django-better-admin-arrayfield = "^1.0.5"

[tool.poetry.scripts]

[tool.poetry.dev-dependencies]
pytest = "^3.4"
black = {version = "^18.3-alpha.0", allow-prereleases = true}
flake8 = "^3.7.9"
django-debug-toolbar = "^2.1"
rope = "^0.16.0"

[tool.black]
exclude ='''
(
  /(
    | \.git
    | \.pytest_cache
    | build
    | dist
    | web/node_modules
    | .*\/migrations
  )/
)
'''

[build-system]
requires = ["poetry==1.0.2"]
build-backend = "poetry.masonry.api"
